requirements.txt
imageio==2.34.0
keras==3.2.1
librosa==0.9.2
loguru==0.6.0
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
Pillow==9.0.1
PyYAML==5.4.1
scikit_learn==0.23.1
scipy==1.8.0
six==1.16.0
SoundFile==0.10.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.0
